feat: ga now can recieve parameters
GA is able to recieve changes of pop_size, genes_num, crossover rate and mutation rate from renderer process live, but changes does not take effect instantly. but as follow:
if new value for pop_size or genes_num is passed, it does not take effect until the next GA_Thread initialization.
if crossover or mutation rate is passed while GA is running and in process of crossover or mutation, the change is going to take effect on the next generation which means that changes of these values in the middle of one generation is not possible.