refactor: data parsing to JSON is now done only on renderer process
renderer process used to pass data to prime & side webviews as row buffers, be parsed in both of them before treated, now the data is parsed in renderer process and passed to the webviews as JSON object, this should improve drawing speed of charts.
fix issue when clicking step forward button twice on the beginning of the GA, it generates only 1 point.
removed unneeded condition checks.