hotfix: fixed cases where GA would block
if GA is heavy enough to take time generating each generation, there is a potential that it's going to block in some cases, specially at start, revise changes comments on ga.py for more.
dropped the deprecated 'setup_finished' handeling case.
blocking the buttons when stop is clicked is moved to disable button after the click and not after the 'stopped' message is recieved, because it can take considerable amount of time for the GA to finish a generation and produce the stopped event.