reading setting is not async anymore
reading settings is switched to be synchronous to avoid running into undefined settings if renderer process of the app is ready before the async settings on the preload (that problem didn't happen on main process so far).
changed closable to be changed on keypress to avoid it being changed by unnecessary pressed keys on textfields.
title to pin buttons is added.