globalize scss for sync-param and add random button compatibility
scss of sync-param is now global in params.scss with minor patches for compatibility on both main and ga-cp.
random button can handle sync type params now.
fix pin button of sync-param to be able to pin and disable the whole param.
other fixes and code refactors to remove redundancy.